# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  displayName: string;
  avatarColor: string;
  createdAt: string;
}

export interface CreateUserBody {
  /**
   * @minLength 1
   * @maxLength 32
   */
  displayName: string;
}

export interface Room {
  id: string;
  name: string;
  description: string;
  emoji: string;
  createdAt: string;
  messageCount: number;
  lastMessageAt?: string | null;
  lastMessagePreview?: string | null;
}

export interface RoomDetail {
  id: string;
  name: string;
  description: string;
  emoji: string;
  createdAt: string;
  messageCount: number;
  memberCount: number;
}

export interface CreateRoomBody {
  /**
   * @minLength 1
   * @maxLength 64
   */
  name: string;
  /** @maxLength 200 */
  description?: string;
  /** @maxLength 8 */
  emoji?: string;
}

export interface Message {
  id: string;
  roomId: string;
  userId: string;
  userDisplayName: string;
  userAvatarColor: string;
  body: string;
  createdAt: string;
}

export interface SendMessageBody {
  userId: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  body: string;
}

export interface DashboardSummary {
  totalRooms: number;
  totalMessages: number;
  totalUsers: number;
  messagesToday: number;
  activeRoomsToday: number;
}

export interface ActivityItem {
  messageId: string;
  roomId: string;
  roomName: string;
  roomEmoji: string;
  userDisplayName: string;
  userAvatarColor: string;
  body: string;
  createdAt: string;
}