# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Chat application API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Chat room operations
  - name: messages
    description: Message operations
  - name: users
    description: User operations
  - name: dashboard
    description: Dashboard summaries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or fetch a user by display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by id
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all chat rooms
      responses:
        "200":
          description: Rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new chat room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomBody"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{roomId}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a single room with details
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetail"

  /rooms/{roomId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /rooms/{roomId}/members:
    get:
      operationId: listRoomMembers
      tags: [rooms]
      summary: List recently active members in a room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Aggregate stats across all rooms
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Recent messages across all rooms
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    User:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        avatarColor:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, displayName, avatarColor, createdAt]

    CreateUserBody:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 32
      required: [displayName]

    Room:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        emoji:
          type: string
        createdAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        lastMessagePreview:
          type: string
          nullable: true
      required: [id, name, description, emoji, createdAt, messageCount]

    RoomDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        emoji:
          type: string
        createdAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        memberCount:
          type: integer
      required: [id, name, description, emoji, createdAt, messageCount, memberCount]

    CreateRoomBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          maxLength: 200
        emoji:
          type: string
          maxLength: 8
      required: [name]

    Message:
      type: object
      properties:
        id:
          type: string
        roomId:
          type: string
        userId:
          type: string
        userDisplayName:
          type: string
        userAvatarColor:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, roomId, userId, userDisplayName, userAvatarColor, body, createdAt]

    SendMessageBody:
      type: object
      properties:
        userId:
          type: string
        body:
          type: string
          minLength: 1
          maxLength: 2000
      required: [userId, body]

    DashboardSummary:
      type: object
      properties:
        totalRooms:
          type: integer
        totalMessages:
          type: integer
        totalUsers:
          type: integer
        messagesToday:
          type: integer
        activeRoomsToday:
          type: integer
      required: [totalRooms, totalMessages, totalUsers, messagesToday, activeRoomsToday]

    ActivityItem:
      type: object
      properties:
        messageId:
          type: string
        roomId:
          type: string
        roomName:
          type: string
        roomEmoji:
          type: string
        userDisplayName:
          type: string
        userAvatarColor:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [messageId, roomId, roomName, roomEmoji, userDisplayName, userAvatarColor, body, createdAt]