# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create or fetch a user by display name
 */
export const createUserBodyDisplayNameMax = 32;

export const CreateUserBody = zod.object({
  displayName: zod.string().min(1).max(createUserBodyDisplayNameMax),
});

export const CreateUserResponse = zod.object({
  id: zod.string(),
  displayName: zod.string(),
  avatarColor: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get a user by id
 */
export const GetUserParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserResponse = zod.object({
  id: zod.string(),
  displayName: zod.string(),
  avatarColor: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all chat rooms
 */
export const ListRoomsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string(),
  emoji: zod.string(),
  createdAt: zod.coerce.date(),
  messageCount: zod.number(),
  lastMessageAt: zod.coerce.date().nullish(),
  lastMessagePreview: zod.string().nullish(),
});
export const ListRoomsResponse = zod.array(ListRoomsResponseItem);

/**
 * @summary Create a new chat room
 */
export const createRoomBodyNameMax = 64;

export const createRoomBodyDescriptionMax = 200;

export const createRoomBodyEmojiMax = 8;

export const CreateRoomBody = zod.object({
  name: zod.string().min(1).max(createRoomBodyNameMax),
  description: zod.string().max(createRoomBodyDescriptionMax).optional(),
  emoji: zod.string().max(createRoomBodyEmojiMax).optional(),
});

/**
 * @summary Get a single room with details
 */
export const GetRoomParams = zod.object({
  roomId: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string(),
  emoji: zod.string(),
  createdAt: zod.coerce.date(),
  messageCount: zod.number(),
  memberCount: zod.number(),
});

/**
 * @summary List messages in a room
 */
export const ListMessagesParams = zod.object({
  roomId: zod.coerce.string(),
});

export const ListMessagesResponseItem = zod.object({
  id: zod.string(),
  roomId: zod.string(),
  userId: zod.string(),
  userDisplayName: zod.string(),
  userAvatarColor: zod.string(),
  body: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Send a message to a room
 */
export const SendMessageParams = zod.object({
  roomId: zod.coerce.string(),
});

export const sendMessageBodyBodyMax = 2000;

export const SendMessageBody = zod.object({
  userId: zod.string(),
  body: zod.string().min(1).max(sendMessageBodyBodyMax),
});

/**
 * @summary List recently active members in a room
 */
export const ListRoomMembersParams = zod.object({
  roomId: zod.coerce.string(),
});

export const ListRoomMembersResponseItem = zod.object({
  id: zod.string(),
  displayName: zod.string(),
  avatarColor: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListRoomMembersResponse = zod.array(ListRoomMembersResponseItem);

/**
 * @summary Aggregate stats across all rooms
 */
export const GetDashboardSummaryResponse = zod.object({
  totalRooms: zod.number(),
  totalMessages: zod.number(),
  totalUsers: zod.number(),
  messagesToday: zod.number(),
  activeRoomsToday: zod.number(),
});

/**
 * @summary Recent messages across all rooms
 */
export const GetRecentActivityResponseItem = zod.object({
  messageId: zod.string(),
  roomId: zod.string(),
  roomName: zod.string(),
  roomEmoji: zod.string(),
  userDisplayName: zod.string(),
  userAvatarColor: zod.string(),
  body: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);