# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  messageId: string;
  roomId: string;
  roomName: string;
  roomEmoji: string;
  userDisplayName: string;
  userAvatarColor: string;
  body: string;
  createdAt: Date;
}