# lib/api-zod/src/generated/types/createRoomBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomBody {
  /**
   * @minLength 1
   * @maxLength 64
   */
  name: string;
  /** @maxLength 200 */
  description?: string;
  /** @maxLength 8 */
  emoji?: string;
}