# lib/api-zod/src/generated/types/createUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserBody {
  /**
   * @minLength 1
   * @maxLength 32
   */
  displayName: string;
}