# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRooms: number;
  totalMessages: number;
  totalUsers: number;
  messagesToday: number;
  activeRoomsToday: number;
}