# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  roomId: string;
  userId: string;
  userDisplayName: string;
  userAvatarColor: string;
  body: string;
  createdAt: Date;
}