# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  name: string;
  description: string;
  emoji: string;
  createdAt: Date;
  messageCount: number;
  lastMessageAt?: Date | null;
  lastMessagePreview?: string | null;
}