# lib/api-zod/src/generated/types/roomDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface RoomDetail {
  id: string;
  name: string;
  description: string;
  emoji: string;
  createdAt: Date;
  messageCount: number;
  memberCount: number;
}