# lib/api-zod/src/generated/types/sendMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chat application API
 * OpenAPI spec version: 0.1.0
 */

export interface SendMessageBody {
  userId: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  body: string;
}